backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-auth>=2.35.0
google-cloud-storage>=2.18.0
google-cloud-firestore>=2.19.0
requests>=2.32.0
Pillow>=10.4.0
scikit-image>=0.25.0
numpy>=1.26.4
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
